assorted fixes
